// src/types/types.ts
export interface IData {
    date: string;
    units: string;
    units_lastyear: string;
    revenue: string;
    revenue_lastyear: string;
    cogs: string;
    ads_cost: string;
    best_sellers_rank: string;
    ppc_revenue_adjusted: string;
    ads_cost_adjusted: string;
    ppc_revenue: string;
    [key: string]: string;
  }